requirements.txt
aiofiles==24.1.0
aiomysql==0.2.0
asgiref==3.8.1
asyncpg==0.30.0
certifi==2024.8.30
charset-normalizer==3.4.0
Django==4.2
django-extensions==3.2.3
djangorestframework==3.15.2
idna==3.10
jsonfield==3.1.0
numpy==2.1.2
pandas==2.2.3
psycopg2-binary==2.9.10
PyMySQL==1.1.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
requests==2.32.3
six==1.16.0
sqlparse==0.5.1
tzdata==2024.2
urllib3==2.2.3